% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Aneufinder.R
\name{Aneufinder}
\alias{Aneufinder}
\title{Wrapper function for the aneufinder package}
\usage{
Aneufinder(inputfolder, outputfolder, config = makeConfig())
}
\arguments{
\item{inputfolder}{Folder with either BAM or BED files.}

\item{outputfolder}{Folder to output the results. If it does not exist it will be created.}

\item{config}{A file with parameters that are used by the various subfunctions. Alternatively a list with parameters created by \code{\link{makeConfig}}.}
}
\description{
This function is an easy-to-use wrapper to \link[aneufinder:binning]{bin the data}, \link[aneufinder:findCNVs]{find copy-number-variations}, \link[aneufinder:findSCEs]{find sister-chromatid-exchange} events and plot the results.
}
\author{
Aaron Taudt
}

