% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBreakpoints.R
\name{annotateBreakpoints}
\alias{annotateBreakpoints}
\title{Annotate breakpoints}
\usage{
annotateBreakpoints(breakpoints)
}
\arguments{
\item{breakpoints}{A \code{\link{GRanges-class}} as returned by \code{\link{getBreakpoints}}.}
}
\value{
The input \code{\link{GRanges-class}} with additinal column 'type'.
}
\description{
Annotate breakpoints as sister-chromatid-exchange (SCE), copy-number-breakpoint (CNB).
}
\examples{
## Get an example BED file with single-cell-sequencing reads
bedfile <- system.file("extdata", "KK150311_VI_07.bam.bed.gz", package="AneuFinderData")
## Bin the data into bin size 1Mp
readfragments <- binReads(bedfile, assembly='mm10', binsize=1e6,
                  chromosomes=c(1:19,'X','Y'), reads.return=TRUE)
binned <- binReads(bedfile, assembly='mm10', binsize=1e6,
                  chromosomes=c(1:19,'X','Y'))
## Fit the Hidden Markov Model
model <- findCNVs.strandseq(binned[[1]])
## Add confidence intervals
breakpoints <- getBreakpoints(model, readfragments)

}
