% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/binning.R
\name{bam2GRanges}
\alias{bam2GRanges}
\title{Import BAM file into GRanges}
\usage{
bam2GRanges(file, bamindex = file, chromosomes = NULL,
  pairedEndReads = FALSE, keep.duplicate.reads = TRUE)
}
\arguments{
\item{pairedEndReads}{Set to \code{TRUE} if you have paired-end reads in your file.}

\item{keep.duplicate.reads}{A logical indicating whether or not duplicate reads should be kept.}
}
\description{
Import aligned reads from a BAM file into a \code{\link{GRanges}} object.
}

