% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convert2GRanges.R
\name{bed2GRanges}
\alias{bed2GRanges}
\title{Read bed-file into GRanges}
\usage{
bed2GRanges(bedfile, skip = 0)
}
\arguments{
\item{bedfile}{Filename of the bed or bed.gz file.}

\item{skip}{Number of lines to skip at the beginning.}
}
\value{
A \code{\link{GRanges}} object with the contents of the bed-file.
}
\description{
This is a simple convenience function to read a (compressed) bed-file into a \code{\link{GRanges}} object. The bed-file is expected to have the following fields: \code{chrom, chromStart, chromEnd, name, score, strand}.
}
\author{
Aaron Taudt
}

