% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findCNVs.R
\name{bi.edivisive.findCNVs}
\alias{bi.edivisive.findCNVs}
\title{Find copy number variations (edivisive, bivariate)}
\usage{
bi.edivisive.findCNVs(binned.data, ID = NULL, min.ground.ploidy = 0.5,
  max.ground.ploidy = 6, R = 10, sig.lvl = 0.1)
}
\arguments{
\item{binned.data}{A \link{GRanges} object with binned read counts.}

\item{ID}{An identifier that will be used to identify this sample in various downstream functions. Could be the file name of the \code{binned.data} for example.}

\item{min.ground.ploidy}{Minimum ground-ploidy state considered. Set to 1.5 for normal data and 0.5 for Strand-seq data.}

\item{max.ground.ploidy}{Maximum ground-ploidy state considered.}

\item{R}{method-edivisive: The maximum number of random permutations to use in each iteration of the permutation test (see \code{\link[ecp]{e.divisive}}). Increase this value to increase accuracy on the cost of speed.}

\item{sig.lvl}{method-edivisive: The level at which to sequentially test if a proposed change point is statistically significant (see \code{\link[ecp]{e.divisive}}). Increase this value to find more breakpoints.}
}
\value{
An \code{\link{aneuHMM}} object.
}
\description{
Classify the binned read counts into several states which represent copy-number-variation. The function uses the \code{\link{e.divisive}} function to segment the genome.
}
