% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findCNVs.R
\name{biDNAcopy.findCNVs}
\alias{biDNAcopy.findCNVs}
\title{Find copy number variations (DNAcopy, bivariate)}
\usage{
biDNAcopy.findCNVs(binned.data, ID = NULL, min.ground.ploidy = 0.5,
  max.ground.ploidy = 6)
}
\arguments{
\item{binned.data}{A \link{GRanges} object with binned read counts.}

\item{ID}{An identifier that will be used to identify this sample in various downstream functions. Could be the file name of the \code{binned.data} for example.}

\item{min.ground.ploidy}{Minimum ground-ploidy state considered. Set to 1.5 for normal data and 0.5 for Strand-seq data.}

\item{max.ground.ploidy}{Maximum ground-ploidy state considered.}
}
\value{
An \code{\link{aneuHMM}} object.
}
\description{
\code{biDNAcopy.findCNVs} classifies the binned read counts into several states which represent copy-number-variation using read count information from both strands.
}
