% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documented_objects.R
\name{binned.data}
\alias{binned.data}
\title{Binned read counts}
\description{
A \link{GRanges} object which contains binned read counts as meta data column \code{reads}. It is output of the various \link{binning} functions.
}
