% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/binning.R
\name{binning}
\alias{bam2binned}
\alias{bed2binned}
\alias{bedGraph2binned}
\alias{binning}
\title{Convert aligned reads from various file formats into read counts in equidistant bins}
\usage{
bam2binned(bamfile, bamindex = bamfile, ID = basename(bamfile),
  pairedEndReads = FALSE, outputfolder = "binned_data", binsizes = NULL,
  reads.per.bin = NULL, numbins = NULL, chromosomes = NULL,
  save.as.RData = FALSE, calc.complexity = TRUE, min.mapq = 10,
  remove.duplicate.reads = TRUE, outputfolder.fragments = NULL,
  return.fragments = FALSE)

bed2binned(bedfile, chrom.length.file, ID = basename(bedfile),
  outputfolder = "binned_data", binsizes = NULL, reads.per.bin = NULL,
  numbins = NULL, chromosomes = NULL, save.as.RData = FALSE,
  calc.complexity = TRUE, min.mapq = 10, remove.duplicate.reads = TRUE,
  outputfolder.fragments = NULL, return.fragments = FALSE)

bedGraph2binned(bedGraphfile, chrom.length.file, ID = basename(bedGraphfile),
  outputfolder = "binned_data", binsizes = NULL, reads.per.bin = NULL,
  numbins = NULL, chromosomes = NULL, save.as.RData = FALSE,
  calc.complexity = TRUE, min.mapq = 10, remove.duplicate.reads = TRUE,
  outputfolder.fragments = NULL, return.fragments = FALSE)
}
\arguments{
\item{bamfile}{A file in BAM format.}

\item{bamindex}{BAM index file. Can be specified without the .bai ending. If the index file does not exist it will be created and a warning is issued.}

\item{ID}{An identifier that will be used to identify the file throughout the workflow and in plotting.}

\item{pairedEndReads}{Set to \code{TRUE} if you have paired-end reads in your file.}

\item{outputfolder}{Folder to which the binned data will be saved. If the specified folder does not exist, it will be created.}

\item{binsizes}{An integer vector with bin sizes. If more than one value is given, output files will be produced for each bin size.}

\item{reads.per.bin}{Approximate number of desired reads per bin. The bin size will be selected accordingly. Output files are produced for each value.}

\item{numbins}{Number of bins per chromosome. Each chromosome will have a different binsize! DO NOT USE UNLESS YOU KNOW WHAT YOU ARE DOING. Output files are produced for each value.}

\item{chromosomes}{If only a subset of the chromosomes should be binned, specify them here.}

\item{save.as.RData}{If set to \code{FALSE}, no output file will be written. Instead, a \link{GenomicRanges} object containing the binned data will be returned. Only the first binsize will be processed in this case.}

\item{calc.complexity}{A logical indicating whether or not to estimate library complexity.}

\item{min.mapq}{Minimum mapping quality when importing from BAM files.}

\item{remove.duplicate.reads}{A logical indicating whether or not duplicate reads should be removed.}

\item{outputfolder.fragments}{Folder to which the read coordinates from the input file will be saved in \code{\link{GRanges}} format. These are the reads as used for the binning (duplicates removed, minimum mapping quality, etc.).}

\item{return.fragments}{If \code{TRUE} no binning is done and instead, read fragments from the input file are returned in \code{\link{GRanges}} format.}

\item{bedfile}{A file in BED format.}

\item{chrom.length.file}{A file which contains the chromosome lengths in basepairs. The first column contains the chromosome name and the second column the length (see also \code{\link{chrom.length.file}}.}

\item{bedGraphfile}{A file in bedGraph format.}
}
\description{
Convert aligned reads in .bam or .bed format into read counts in equidistant windows. Convert signal values in .bedGraph format to signal counts in equidistant windows.
}
\details{
Convert aligned reads or signal values from various file formats into read counts in equidistant windows (bins). bam2binned() and bed2binned() use 'GenomicRanges::countOverlaps()' to calculate the read counts. Therefore, with small binsizes and large read lengths, one read fragment will often overlap more than one bin.
bedGraph2binned() sets the maximum signal value in a bin as value for that bin.
}
\section{Functions}{
\itemize{
\item \code{bam2binned}: Bin reads in BAM format

\item \code{bed2binned}: Bin reads in BED format

\item \code{bedGraph2binned}: Bin reads in bedGraph format
}}
\examples{
## Get an example BAM file with single-cell-sequencing reads
bamfile <- system.file("extdata/BB140820_I_002.bam", package="aneufinder")
## Bin the BAM file into bin size 200000bp
binned.data <- bam2binned(bamfile, binsize=200000, chromosomes=c(1:22,'X','Y'), save.as.RData=FALSE)
}
\author{
Aaron Taudt
}

