% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/findSCEs.R
\name{bivariate.findSCEs}
\alias{bivariate.findSCEs}
\title{Find copy number variations (bivariate)}
\usage{
bivariate.findSCEs(binned.data, ID, eps = 0.001, init = "standard",
  max.time = -1, max.iter = -1, num.trials = 1, eps.try = NULL,
  num.threads = 1, read.cutoff.quantile = 0.999, GC.correction = TRUE,
  allow.odd.states = TRUE, states = c("zero-inflation", "nullsomy",
  "monosomy", "disomy", "trisomy", "tetrasomy", "multisomy"))
}
\arguments{
\item{allow.odd.states}{If set to \code{TRUE}, all states are allowed. If \code{FALSE}, only states which have an even multiplicity (plus nullsomy-monosomy states) are allowed, e.g. disomy-disomy, monosomy-trisomy.}
}
\description{
\code{bivariate.findSCEs} finds CNVs using read count information from both strands.
}

