% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/chrom.length.file.R
\name{chrom.length.file}
\alias{chrom.length.file}
\title{Format of the \code{chrom.length.file}}
\description{
The \code{chrom.length.file} can be used in the following \pkg{\link{aneufinder}} functions: \code{\link{bed2binned}}, \code{\link{bedGraph2binned}}.
}
\details{
In the first column, it contains the chromosome names and in the second column, it contains the chromosome lengths in basepairs. The file has no header. Example:
\tabular{ll}{
chr1 \tab 249250621 \cr
chr2 \tab 243199373 \cr
chr3 \tab 198022430 \cr
chr4 \tab 191154276 \cr
... \tab ... \cr
}
}

