% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qualityMeasures.R
\name{clusterByQuality}
\alias{clusterByQuality}
\title{Cluster based on quality variables}
\usage{
clusterByQuality(hmms, G = NULL)
}
\description{
This function uses the \pkg{\link{mclust}} package to cluster the input samples based on various quality measures.
}
\details{
The employed quality measures are:
\itemize{
\item Spikyness
\item Entropy
\item Number of segments
\item Bhattacharrya distance
}
}
\author{
Aaron Taudt
}

