% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterHMMs.R
\name{clusterHMMs}
\alias{clusterHMMs}
\title{Cluster objects}
\usage{
clusterHMMs(hmms, cluster = TRUE, classes = NULL, exclude.regions = NULL)
}
\arguments{
\item{hmms}{A list of \code{\link{aneuHMM}} or \code{\link{aneuBiHMM}} objects or a character vector of files that contains such objects.}

\item{cluster}{Either \code{TRUE} or \code{FALSE}, indicating whether the samples should be clustered by similarity in their CNV-state.}

\item{classes}{A vector with class labels the same length as \code{hmms}. If supplied, the clustering will be ordered optimally with respect to the class labels (see \code{\link[ReorderCluster]{RearrangeJoseph}}).}

\item{exclude.regions}{A \code{\link{GRanges}} with regions that will be excluded from the computation of the clustering. This can be useful to exclude regions with artifacts.}
}
\value{
An list() with ordered ID indices and the hierarchical clustering.
}
\description{
Cluster a list of \code{\link{aneuHMM}} or \code{\link{aneuBiHMM}} objects by similarity in their CNV-state.
}
