% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{colors}
\alias{breakpointColors}
\alias{colors}
\alias{stateColors}
\alias{strandColors}
\title{\pkg{AneuFinder} color scheme}
\usage{
stateColors(states = c("zero-inflation", paste0(0:10, "-somy"), "total"))

strandColors(strands = c("+", "-"))

breakpointColors(breaktypes = c("CNB", "SCE", "CNB+SCE", "other"))
}
\arguments{
\item{states}{A character vector with states whose color should be returned.}

\item{strands}{A character vector with strands whose color should be returned. Any combination of \code{c('+','-','*')}.}

\item{type}{A character vector with breakpoint types whose color should be returned. Any combination of \code{c('CNB','SCE','CNB+SCE','other')}.}
}
\value{
A character vector with colors.
}
\description{
Get the color schemes that are used in the AneuFinder plots.
}
\section{Functions}{
\itemize{
\item \code{stateColors}: Colors that are used for the states.

\item \code{strandColors}: Colors that are used to distinguish strands.

\item \code{breakpointColors}: Colors that are used for breakpoint types.
}}
\examples{
## Make a nice pie chart with the AneuFinder state color scheme
statecolors <- stateColors()
pie(rep(1,length(statecolors)), labels=names(statecolors), col=statecolors)

## Make a nice pie chart with the AneuFinder strand color scheme
strandcolors <- strandColors()
pie(rep(1,length(strandcolors)), labels=names(strandcolors), col=strandcolors)

## Make a nice pie chart with the AneuFinder breakpoint-type color scheme
breakpointcolors <- breakpointColors()
pie(rep(1,length(breakpointcolors)), labels=names(breakpointcolors), col=breakpointcolors)

}

