% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareMethods.R
\name{compareModels}
\alias{compareModels}
\title{Compare copy number models}
\usage{
compareModels(model1, model2)
}
\arguments{
\item{model1}{An \code{\link{aneuHMM}} object or file that contains such an object.}

\item{model2}{An \code{\link{aneuHMM}} object or file that contains such an object.}
}
\value{
A numeric.
}
\description{
Compare two \code{\link{aneuHMM}} objects. The function computes the fraction of copy number calls that is concordant between both models.
}
\author{
Aaron Taudt
}
