% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/refine_changepoints.R
\name{concGRangesInRange}
\alias{concGRangesInRange}
\title{Concatenate GRanges in range}
\usage{
concGRangesInRange(grlist, chromosome, start, end)
}
\arguments{
\item{grlist}{A list with \code{\link{GRanges}} objects or files that contain such objects.}

\item{chromosome,start,end}{Coordinates of the range.}
}
\description{
Concatenate \code{\link{GRanges}} objects in range.
}

