% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/correctionMethods.R
\name{correctMappability}
\alias{correctMappability}
\title{Mappability correction}
\usage{
correctMappability(binned.data.list)
}
\arguments{
\item{binned.data.list}{A \code{list()} with \code{\link{binned.data}} objects or a list of filenames containing such objects.}
}
\description{
Correct a list of \code{\link{binned.data}} by mappability. WARNING: This approach works only for a set of super heterogeneous samples.
}
\author{
Aaron Taudt
}

