% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findCNVs.R
\name{DNAcopy.findCNVs}
\alias{DNAcopy.findCNVs}
\title{Find copy number variations (DNAcopy)}
\usage{
DNAcopy.findCNVs(binned.data, ID = NULL, most.frequent.state = "2-somy",
  count.cutoff.quantile = 0.999, strand = "*")
}
\arguments{
\item{binned.data}{A \link{GRanges} object with binned read counts.}

\item{ID}{An identifier that will be used to identify this sample in various downstream functions. Could be the file name of the \code{binned.data} for example.}

\item{most.frequent.state}{One of the states that were given in \code{states} or \code{NULL}. The specified state is assumed to be the most frequent one. This can help the fitting procedure to converge into the correct fit.}

\item{count.cutoff.quantile}{A quantile between 0 and 1. Should be near 1. Read counts above this quantile will be set to the read count specified by this quantile. Filtering very high read counts increases the performance of the Baum-Welch fitting procedure. However, if your data contains very few peaks they might be filtered out. Set \code{count.cutoff.quantile=1} in this case.}

\item{strand}{Run the HMM only for the specified strand. One of \code{c('+', '-', '*')}.}
}
\value{
An \code{\link{aneuHMM}} object.
}
\description{
\code{findCNVs} classifies the binned read counts into several states which represent copy-number-variation.
}

