% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/easyAneufinder.R
\name{easyAneufinder}
\alias{easyAneufinder}
\title{Wrapper function for the aneufinder package}
\usage{
easyAneufinder(inputfolder, configfile, outputfolder)
}
\arguments{
\item{inputfolder}{Folder with BAM files.}

\item{configfile}{A file with parameters that are used by the various functions.}

\item{outputfolder}{Folder to put the results. If it does not exist it will be created.}
}
\description{
This function is an easy-to-use wrapper to \link[aneufinder:binning]{bin the data}, \link[aneufinder:findCNVs]{find copy-number-variations}, \link[aneufinder:findSCEs]{find sister-chromatid-exchange} events and plot the results.
}
\author{
Aaron Taudt
}

