% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/findSCEs.R
\name{filterSegments}
\alias{filterSegments}
\title{Filter segments by minimal size}
\usage{
filterSegments(model, min.seg.width.binsize = 2, min.seg.width.bp = NULL)
}
\arguments{
\item{model}{A \code{\link{aneuHMM}} or \code{\link{aneuBiHMM}} object.}

\item{min.seg.width}{The minimum segment width in base-pairs. The default \code{-1} will be twice the bin size of the given \code{model}.}
}
\value{
The input \code{model} with adjusted segments.
}
\description{
\code{filterSegments} filters out segments below a specified minimal segment size. This can be useful to get rid of boundary effects from the Hidden Markov approach.
}
\author{
Aaron Taudt
}

