% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findBreakpoints.R
\name{findBreakpoints}
\alias{findBreakpoints}
\title{Find breakpoints}
\usage{
findBreakpoints(model, fragments, breakpoint.quantile = 0.99)
}
\arguments{
\item{model}{An \code{\link{aneuBiHMM}} object or a file that contains such an object.}

\item{fragments}{A \code{\link[GenomicRanges]{GRanges}} object with read fragments.}

\item{breakpoint.quantile}{A quantile cutoff between 0 and 1 for breakpoint detection. Higher values will result in higher precision but lower sensitivity.}
}
\description{
Breakpoint detection is done via a dynamic windowing approach on read resolution.
}
\author{
Aaron Taudt, David Porubsky, Ashley Sanders
}

