% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getChromLengths.R
\name{getChromLengths}
\alias{getChromLengths}
\title{Chromosome lengths for different assemblies}
\usage{
getChromLengths(assembly = "hg19")
}
\arguments{
\item{assembly}{One of the following: \code{c('hg19','hg18','mm10','mm9','rn4')}.}
}
\value{
A named vector with chromosome lengths.
}
\description{
Get the chromosome lengths of different common assemblies. Only the standard chromosomes (no chrN_random) will be returned.
}

