% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualityMeasures.R
\name{getQC}
\alias{getQC}
\title{Obtain a data.frame with quality metrics}
\usage{
getQC(hmms)
}
\arguments{
\item{hmms}{A list of \code{\link{aneuHMM}} objects or a list of files that contain such objects.}
}
\value{
A data.frame with columns
}
\description{
Obtain a data.frame with quality metrics from a list of \code{\link{aneuHMM}} objects or a list of files that contain such objects.
}
\author{
Aaron Taudt
}

