% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSCEs.R
\name{getSCEcoordinates}
\alias{getSCEcoordinates}
\title{Get SCE coordinates}
\usage{
getSCEcoordinates(model, resolution = c(3, 6), min.segwidth = 2,
  fragments = NULL, min.reads = 50)
}
\arguments{
\item{model}{An \code{\link{aneuBiHMM}} object.}

\item{resolution}{An integer vector specifying the resolution at bin level at which to scan for SCE events.}

\item{min.segwidth}{Minimum segment length in bins when scanning for SCE events.}

\item{fragments}{A \code{\link{GRanges}} object with read fragments or a file that contains such an object. These reads will be used for fine mapping of the SCE events.}

\item{min.reads}{Minimum number of reads required for SCE refinement.}
}
\value{
A \code{\link{GRanges}} object containing the SCE coordinates.
}
\description{
Extracts the coordinates of a sister chromatid exchanges (SCE) from an \code{\link{aneuBiHMM}} object.
}
\examples{
## Get an example BED file with single-cell-sequencing reads
bedfile <- system.file("extdata/KK150311-VI_07.bam.bed.gz", package="aneufinder")
## Bin the BAM file into bin size 1Mp
binned <- binReads(bedfile, format='bed', assembly='hg19', binsize=1e6,
                  chromosomes=c(1:22,'X','Y'), pairedEndReads=TRUE)
## Fit the Hidden Markov Model
model <- findSCEs(binned[[1]], eps=0.1, max.time=60)
## Find sister chromatid exchanges
model$sce <- getSCEcoordinates(model)
print(model$sce)
plot(model)

}
\author{
Aaron Taudt
}

