% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{heatmapAneuploidies}
\alias{heatmapAneuploidies}
\title{Plot aneuploidy state}
\usage{
heatmapAneuploidies(hmms, ylabels = NULL, cluster = TRUE,
  as.data.frame = FALSE)
}
\arguments{
\item{hmms}{A list of \code{\link{aneuHMM}} objects or a character vector with files that contain such objects.}

\item{ylabels}{A vector with labels for the y-axis. The vector must have the same length as \code{hmms}. If \code{NULL} the IDs from the \code{\link{aneuHMM}} objects will be used.}

\item{cluster}{If \code{TRUE}, the samples will be clustered by similarity in their CNV-state.}

\item{as.data.frame}{If \code{TRUE}, instead of a plot, a data.frame with the aneuploidy state for each sample will be returned.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot}} object or a data.frame, depending on option \code{as.data.frame}.
}
\description{
Plot a heatmap of aneuploidy state for multiple samples. Samples can be clustered and the output can be returned as data.frame.
}
\examples{
## Get results from a small-cell-lung-cancer
folder <- system.file("extdata", "primary-lung", "hmms", package="AneuFinderData")
files <- list.files(folder, full.names=TRUE)
## Plot the ploidy state per chromosome
heatmapAneuploidies(files, cluster=FALSE)
## Return the ploidy state as data.frame
df <- heatmapAneuploidies(files, cluster=FALSE, as.data.frame=TRUE)
head(df)

}
\author{
Aaron Taudt
}
