% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{heatmapGenomewide}
\alias{heatmapGenomewide}
\title{Genome wide heatmap of CNV-state}
\usage{
heatmapGenomewide(hmm.list, ylabels = NULL, file = NULL, cluster = TRUE,
  plot.SCE = TRUE, hotspots = NULL)
}
\arguments{
\item{hmm.list}{A list of \code{\link{aneuHMM}} objects or files that contain such objects.}

\item{ylabels}{A vector with labels for the y-axis. The vector must have the same length as \code{hmm.list}. If \code{NULL} the IDs from the \code{\link{aneuHMM}} objects will be used.}

\item{file}{A PDF file to which the heatmap will be plotted.}

\item{cluster}{Either \code{TRUE} or \code{FALSE}, indicating whether the samples should be clustered by similarity in their CNV-state.}

\item{plot.SCE}{Logical indicating whether SCE events should be plotted.}

\item{hotspots}{A \code{\link{GRanges}} object with coordinates of genomic hotspots (see \code{\link{hotspotter}}).}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot}} object or \code{NULL} if a file was specified.
}
\description{
Plot a genome wide heatmap of copy number variation state. This heatmap is best plotted to file, because in most cases it will be too big for cleanly plotting it to screen.
}

