% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotspotter.R
\name{hotspotter.variable}
\alias{hotspotter.variable}
\title{Find hotspots of genomic events}
\usage{
hotspotter.variable(breakpoints, confint, pval = 0.05, spacing.bp = 5000)
}
\arguments{
\item{breakpoints}{A list with \code{\link{GRanges-class}} object containing the coordinates of the genomic events and their confidence intervals.}

\item{confint}{Confidence interval that was used for breakpoint estimation.}

\item{pval}{P-value cutoff for hotspots.}

\item{spacing.bp}{Spacing of datapoints for KDE in basepairs.}
}
\value{
A list of \code{\link{GRanges-class}} objects containing 1) coordinates of hotspots and 2) p-values within the hotspot.
}
\description{
Find hotspots of genomic events by using kernel density estimation.
}
\details{
The hotspotter uses a gaussian kernel with variable bandwidth to perform a KDE. The bandwidth depends on the confidence intervals of the breakpoints. A p-value is calculated by comparing the density profile of the genomic events with the density profile of a randomly subsampled set of genomic events (bootstrapping).
}
\author{
Aaron Taudt
}
