% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/findCNVs.R, R/findSCEs.R
\name{initializeStates}
\alias{initializeStates}
\title{Initialize state factor levels and distributions}
\usage{
initializeStates(states)

initializeStates(states)
}
\arguments{
\item{states}{A subset of \code{c("zero-inflation","monosomy","disomy","trisomy","tetrasomy","multisomy")}.}

\item{states}{A subset of \code{c("zero-inflation","nullsomy","monosomy","disomy","trisomy","tetrasomy","multisomy")}.}
}
\description{
Initialize the state factor levels and distributions for the specified states.

Initialize the state factor levels and distributions for the specified states.
}

