% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loadModelsFromFiles.R
\name{loadBinnedFromFiles}
\alias{loadBinnedFromFiles}
\title{Load binned data from files}
\usage{
loadBinnedFromFiles(binned.data.list)
}
\arguments{
\item{binned.data.list}{A list of files that contain \code{\link{binned.data}} objects.}
}
\value{
A list() containing all loaded \code{\link{binned.data}} objects.
}
\description{
Load \code{\link{binned.data}} objects from file into a list.
}
\author{
Aaron Taudt
}

