% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFromFiles.R
\name{loadFromFiles}
\alias{loadFromFiles}
\title{Load \pkg{AneuFinder} objects from file}
\usage{
loadFromFiles(files, check.class = c("GRanges", "aneuHMM", "aneuBiHMM"))
}
\arguments{
\item{files}{A list of \code{\link{GRanges}}, \code{\link{aneuHMM}} or \code{\link{aneuBiHMM}} objects or a vector of files that contain such objects.}

\item{check.class}{Any combination of \code{c('GRanges', 'aneuHMM', 'aneuBiHMM')}. If any of the loaded objects does not belong to the specified class, an error is thrown.}
}
\value{
A list of \code{\link{GRanges}}, \code{\link{aneuHMM}} or \code{\link{aneuBiHMM}} objects.
}
\description{
Wrapper to load \pkg{\link{AneuFinder}} objects from file and check the class of the loaded objects.
}
\examples{
## Get some files that you want to load
folder <- system.file("extdata", "primary-lung", "hmms", package="AneuFinderData")
files <- list.files(folder, full.names=TRUE)
## Load and plot the first then
hmms <- loadFromFiles(files[1:10])
lapply(hmms, plot, type='profile')

}

