% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loadModelsFromFiles.R
\name{loadGRangesFromFiles}
\alias{loadGRangesFromFiles}
\title{Load GRanges from files}
\usage{
loadGRangesFromFiles(files)
}
\arguments{
\item{files}{A list of files that contain \code{\link{GRanges}} objects.}
}
\value{
A list() containing all loaded \code{\link{GRanges}} objects.
}
\description{
Load \code{\link{GRanges}} objects from file into a list.
}
\author{
Aaron Taudt
}

