% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadModelsFromFiles.R
\name{loadHmmsFromFiles}
\alias{loadHmmsFromFiles}
\title{Load HMMs from files}
\usage{
loadHmmsFromFiles(hmms, strict = TRUE)
}
\arguments{
\item{hmms}{A list of files that contain \code{\link{aneuHMM}} objects.}

\item{strict}{If any of the loaded objects is not a \code{\link{aneuHMM}} object, an error (\code{strict=TRUE}) or a warning (\code{strict=FALSE}) will be generated.}
}
\value{
A list() containing all loaded \code{\link{aneuHMM}} objects.
}
\description{
Load \code{\link{aneuHMM}} objects from file into a list.
}
\examples{
## Get some files that you want to load
folder <- system.file("extdata/primary-lung/results_univariate", package="aneufinder")
files <- list.files(folder, full.names=TRUE)
## Load and plot the first then
hmms <- loadHmmsFromFiles(files[1:10])
lapply(hmms, plot, type='arrayCGH')

}
\author{
Aaron Taudt
}

