% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeConfig.R
\name{makeConfig}
\alias{makeConfig}
\title{Aneufinder configuration}
\usage{
makeConfig(General = list(reuse.existing.files = TRUE, numCPU = 1),
  Binning = list(binsizes = 5e+05), Correction = list(),
  CNV = list(findCNVs = TRUE), SCE = list(findSCEs = FALSE),
  Plotting = list())
}
\arguments{
\item{General}{General options.}

\item{Binning}{Parameters for binning. See \code{\link{binning}} for which options are available.}

\item{Correction}{Parameters for correction methods.}

\item{CNV}{Parameters for CNV detection. See \code{\link{findCNVs}} for which options are available.}

\item{SCE}{Parameters for SCE detection. See \code{\link{findSCEs}} for which options are available.}

\item{Plotting}{Plotting parameters.}
}
\description{
Make a list of configuration parameters for the \code{\link{Aneufinder}} function.
}
\author{
Aaron Taudt
}

