% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blacklist.R
\name{mergeStrandseqFiles}
\alias{mergeStrandseqFiles}
\title{Merge Strand-seq libraries}
\usage{
mergeStrandseqFiles(files, assembly, chromosomes = NULL,
  pairedEndReads = FALSE, min.mapq = 10, remove.duplicate.reads = TRUE,
  max.fragment.width = 1000)
}
\arguments{
\item{files}{A character vector with files with aligned reads.}

\item{assembly}{Please see \code{\link[GenomeInfoDb]{fetchExtendedChromInfoFromUCSC}} for available assemblies. Only necessary when importing BED files. BAM files are handled automatically. Alternatively a data.frame with columns 'chromosome' and 'length'.}

\item{chromosomes}{If only a subset of the chromosomes should be imported, specify them here.}

\item{pairedEndReads}{Set to \code{TRUE} if you have paired-end reads in your BAM files (not implemented for BED files).}

\item{min.mapq}{Minimum mapping quality when importing from BAM files. Set \code{min.mapq=NULL} to keep all reads.}

\item{remove.duplicate.reads}{A logical indicating whether or not duplicate reads should be removed.}

\item{max.fragment.width}{Maximum allowed fragment length. This is to filter out erroneously wrong fragments due to mapping errors of paired end reads.}
}
\value{
A \code{\link{GRanges}} object with reads.
}
\description{
Merge strand libraries to generate a high-coverage Strand-seq library.
}
\examples{
#files <- list.files('~/work_ERIBA/test/aneufinder_files/DH161028_WT', full.names = TRUE)
#reads <- mergeStrandseqFiles(files, assembly='hg38')
}

