% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.character}
\alias{plot.character}
\title{Plotting function for saved \pkg{\link{AneuFinder}} objects}
\usage{
\method{plot}{character}(x, ...)
}
\arguments{
\item{x}{A filename that contains either \code{\link{binned.data}} or a \code{\link{aneuHMM}}.}

\item{...}{Additional arguments.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot}} object.
}
\description{
Convenience function that loads and plots a \pkg{\link{AneuFinder}} object in one step.
}

