% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.GRanges}
\alias{plot.GRanges}
\title{Plotting function for binned read counts}
\usage{
\method{plot}{GRanges}(x, type = "karyogram", ...)
}
\arguments{
\item{x}{A \code{\link{GRanges}} object with binned read counts.}

\item{type}{Type of the plot, one of \code{c('karyogram', 'histogram', 'arrayCGH')}. You can also specify the type with an integer number.
\describe{
  \item{\code{karyogram}}{A karyogram-like chromosome overview with read counts.}
  \item{\code{histogram}}{A histogram of read counts.}
  \item{\code{arrayCGH}}{An arrayCGH-like chromosome overview with read counts.}
}}

\item{...}{Additional arguments for the different plot types.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot}} object.
}
\description{
Make plots for binned read counts from \code{\link{binned.data}}.
}

