% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.GRangesList}
\alias{plot.GRangesList}
\title{Plotting function for binned read counts (list)}
\usage{
\method{plot}{GRangesList}(x, type = "profile", ...)
}
\arguments{
\item{x}{A \code{\link{GRangesList}} object with binned read counts.}

\item{type}{Type of the plot, one of \code{c('profile', 'histogram', 'karyogram')}. You can also specify the type with an integer number.
\describe{
  \item{\code{karyogram}}{A karyogram-like chromosome overview with read counts.}
  \item{\code{histogram}}{A histogram of read counts.}
  \item{\code{profile}}{An profile with read counts.}
}}

\item{...}{Additional arguments for the different plot types.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot}} object.
}
\description{
Make plots for binned read counts (list) from \code{\link{binned.data}}.
}
