% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotBinnedDataHistogram}
\alias{plotBinnedDataHistogram}
\title{Plot a histogram of binned read counts}
\usage{
plotBinnedDataHistogram(binned.data, strand = "*", chromosome = NULL,
  start = NULL, end = NULL)
}
\arguments{
\item{binned.data}{A \code{\link{GRanges}} object containing binned read counts in meta-column 'counts'.}

\item{strand}{One of c('+','-','*'). Plot the histogram only for the specified strand.}

\item{chromosome, start, end}{Plot the histogram only for the specified chromosome, start and end position.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot}} object.
}
\description{
Plot a histogram of binned read counts from \code{\link{binned.data}}
}

