% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotChromosomes}
\alias{plotChromosomes}
\title{Plot chromosome overview}
\usage{
plotChromosomes(model, file = NULL)
}
\arguments{
\item{model}{A \code{\link{aneuHMM}} object or \code{\link{binned.data}}.}

\item{file}{A PDF file where the plot will be saved.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot}} object or \code{NULL} if a file was specified.
}
\description{
Plot an overview over all the chromosomes with read counts and their CNV-state from a \code{\link{aneuHMM}} object or \code{\link{binned.data}}.
}

