% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotHistogram}
\alias{plotHistogram}
\title{Plot a histogram of binned read counts with fitted mixture distribution}
\usage{
plotHistogram(model)
}
\arguments{
\item{model}{A \code{\link{aneuHMM}} object.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot}} object.
}
\description{
Plot a histogram of binned read counts from with fitted mixture distributions from a \code{\link{aneuHMM}} object.
}
