% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotPCA}
\alias{plotPCA}
\title{Perform a PCA for copy number profiles}
\usage{
plotPCA(hmms, PC1 = 1, PC2 = 2, colorBy = NULL, plot = TRUE)
}
\arguments{
\item{hmms}{A list of \code{\link{aneuHMM}} objects or a character vector with files that contain such objects.}

\item{PC1}{Integer specifying the first of the principal components to plot.}

\item{PC2}{Integer specifying the second of the principal components to plot.}

\item{colorBy}{A character vector of the same length as \code{hmms} which is used to color the points in the plot.}

\item{plot}{Set to \code{FALSE} if you want to return the data.frame that is used for plotting instead of the plot.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot}} object or a data.frame if \code{plot=FALSE}.
}
\description{
Perform a PCA for copy number profiles in \code{\link{aneuHMM}} objects.
}
\examples{
## Get results from a small-cell-lung-cancer
lung.folder <- system.file("extdata", "primary-lung", "hmms", package="AneuFinderData")
lung.files <- list.files(lung.folder, full.names=TRUE)
## Get results from the liver metastasis of the same patient
liver.folder <- system.file("extdata", "metastasis-liver", "hmms", package="AneuFinderData")
liver.files <- list.files(liver.folder, full.names=TRUE)
## Plot a clustered heatmap
classes <- c(rep('lung', length(lung.files)), rep('liver', length(liver.files)))
labels <- c(paste('lung',1:length(lung.files)), paste('liver',1:length(liver.files)))
plotPCA(c(lung.files, liver.files), colorBy=classes, PC1=2, PC2=4)

}

