% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotUnivariateHistogram}
\alias{plotUnivariateHistogram}
\title{Plot a histogram of binned read counts with fitted mixture distribution}
\usage{
plotUnivariateHistogram(model, state = NULL, strand = "*",
  chromosome = NULL, start = NULL, end = NULL)
}
\arguments{
\item{model}{A \code{\link{aneuHMM}} object.}

\item{state}{Plot the histogram only for the specified CNV-state.}

\item{chromosome,start,end}{Plot the histogram only for the specified chromosome, start and end position.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot}} object.
}
\description{
Plot a histogram of binned read counts from with fitted mixture distributions from a \code{\link{aneuHMM}} object.
}

