% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printMethods.R
\name{print.aneuBiHMM}
\alias{print.aneuBiHMM}
\title{Print aneuBiHMM object}
\usage{
\method{print}{aneuBiHMM}(x, ...)
}
\arguments{
\item{x}{An \code{\link{aneuBiHMM}} object.}

\item{...}{Ignored.}
}
\value{
An invisible \code{NULL}.
}
\description{
Print aneuBiHMM object
}

