% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualityMeasures.R
\name{qualityControl}
\alias{qc.bhattacharyya}
\alias{qc.entropy}
\alias{qc.spikyness}
\alias{qualityControl}
\title{Quality control measures for binned read counts}
\usage{
qc.spikyness(counts)

qc.entropy(counts)

qc.bhattacharyya(hmm)
}
\arguments{
\item{counts}{A vector of binned read counts.}

\item{hmm}{An \code{\link{aneuHMM}} object.}
}
\value{
A numeric.
}
\description{
Calculate various quality control measures on binned read counts.
}
\details{
The Shannon entropy is defined as
\eqn{S = - sum( n * log(n) )}, where \eqn{n = counts/sum(counts)}.\cr\cr 
Spikyness is defined as \eqn{K = sum(abs(diff(counts))) / sum(counts)}.
}
\section{Functions}{
\itemize{
\item \code{qc.spikyness}: Calculate the spikyness of a library

\item \code{qc.entropy}: Calculate the Shannon entropy of a library

\item \code{qc.bhattacharyya}: Calculate the Bhattacharyya distance between the '1-somy' and '2-somy' distribution
}}
\author{
Aaron Taudt
}

