% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwConfig.R
\name{readConfig}
\alias{readConfig}
\title{Read aneufinder configuration file}
\usage{
readConfig(configfile)
}
\arguments{
\item{configfile}{Path to the configuration file}
}
\value{
A \code{list} with one entry for each element in \code{configfile}.
}
\description{
Read an aneufinder configuration file into a list structure. The configuration file has to be specified in INI format. R expressions can be used and will be evaluated.
}
\author{
Aaron Taudt
}

