% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBreakpoints.R
\name{refineBreakpoints}
\alias{refineBreakpoints}
\title{Refine breakpoints}
\usage{
refineBreakpoints(model, breakpoints = model$breakpoints, fragments,
  conf = 0.99)
}
\arguments{
\item{model}{An \code{\link{aneuBiHMM}} object or a file that contains such an object.}

\item{breakpoints}{A \code{\link{GRanges}} object with breakpoints and confidence intervals, as returned by function \code{\link{getBreakpoints}}.}

\item{fragments}{A \code{\link{GRanges}} object with read fragments.}

\item{conf}{Desired confidence interval.}
}
\value{
A \code{\link{GRanges}} with breakpoint coordinates and confidence interals.
}
\description{
Refine breakpoints with confidence intervals from an initial estimate (from \code{\link{getBreakpoints}}).
}
\details{
Breakpoints are refined by shifting the breakpoint within its initial confidence interval read by read and maximizing the probability of observing the left-right read distribution.
}
\examples{
## Get an example BED file with single-cell-sequencing reads
bedfile <- system.file("extdata", "KK150311_VI_07.bam.bed.gz", package="AneuFinderData")
## Bin the data into bin size 1Mp
readfragments <- binReads(bedfile, assembly='mm10', binsize=1e6,
                  chromosomes=c(1:19,'X','Y'), reads.return=TRUE)
binned <- binReads(bedfile, assembly='mm10', binsize=1e6,
                  chromosomes=c(1:19,'X','Y'))
## Fit the Hidden Markov Model
model <- findCNVs.strandseq(binned[[1]], eps=0.1, max.time=60)
## Add confidence intervals
breaks <- getBreakpoints(model, readfragments)
## Refine breakpoints
rfbreaks <- refineBreakpoints(model, breakpoints=breaks, fragments=readfragments)

}

