% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/refine_changepoints.R
\name{smoothBinning}
\alias{smoothBinning}
\title{Smooth binning of read fragments}
\usage{
smoothBinning(gr, binsize, stepsize = binsize/10, start = NULL,
  end = NULL)
}
\arguments{
\item{gr}{A \code{\link{GRanges}} object. The GRanges is assumed to have only data for one chromosome.}

\item{binsize}{The size in base-pairs of the fixed-length window.}

\item{stepsize}{The number of base-pairs for sliding the window.}

\item{start,end}{Coordinates of the binning region. If \code{NULL} coordinates will be determined from the data.}
}
\description{
The function bins read fragments with a fixed-length sliding window.
}

