% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/refine_changepoints.R
\name{subsetByCNVprofile}
\alias{subsetByCNVprofile}
\title{Get IDs of a subset of models}
\usage{
subsetByCNVprofile(hmms, profile)
}
\arguments{
\item{hmms}{A list of \code{\link{aneuHMM}} objects or files that contain such objects.}

\item{profile}{A \code{\link{GRanges}} object with metadata column 'expected.state' and optionally columns 'expected.mstate' and 'expected.pstate'.}
}
\value{
A vector with the IDs of the models that are concordant with the given \code{profile}.
}
\description{
Get the IDs of models that have a certain CNV profile.
}

