% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{transCoord}
\alias{transCoord}
\title{Transform genomic coordinates}
\usage{
transCoord(gr)
}
\arguments{
\item{gr}{A \code{\link{GRanges}} object.}
}
\value{
The input \code{\link{GRanges}} with two additional metadata columns 'start.genome' and 'end.genome'.
}
\description{
Add two columns with transformed genomic coordinates to the \code{\link{GRanges}} object. This is useful for making genomewide plots.
}

