% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeBins.R
\name{variableWidthBins}
\alias{variableWidthBins}
\title{Make variable-width bins}
\usage{
variableWidthBins(reads, binsizes, chromosomes = NULL, assembly = NULL)
}
\arguments{
\item{reads}{A \code{\link{GRanges}} with reads. See \code{\link{bam2GRanges}} and \code{\link{bed2GRanges}}.}

\item{binsizes}{A vector with binsizes. Resulting bins will be close to the specified binsizes.}

\item{chromosomes}{A subset of chromosomes for which the bins are generated.}

\item{assembly}{An assembly from which the chromosome lengths are determined. Please see \code{\link[GenomeInfoDb]{fetchExtendedChromInfoFromUCSC}} for available assemblies. This option is ignored if \code{bamfile} is specified. Alternatively a data.frame generated by \code{\link[GenomeInfoDb]{fetchExtendedChromInfoFromUCSC}}.}
}
\value{
A \code{list()} of \code{\link{GRanges}} objects with variable-width bins.
}
\description{
Make variable-width bins based on a reference BAM file. This can be a simulated file (produced by \code{\link{simulateReads}} and aligned with your favourite aligner) or a real reference.
}
\details{
Variable-width bins are produced by first binning the reference BAM file with fixed-width bins and selecting the desired number of reads per bin as the (non-zero) maximum of the histogram. A new set of bins is then generated such that every bin contains the desired number of reads.
}
\author{
Aaron Taudt
}

