% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwConfig.R
\name{writeConfig}
\alias{writeConfig}
\title{Write AneuFinder configuration file}
\usage{
writeConfig(conf, configfile)
}
\arguments{
\item{conf}{A list structure with parameter values. Each entry will be written in one line.}

\item{configfile}{Filename of the outputfile.}
}
\value{
\code{NULL}
}
\description{
Write an AneuFinder configuration file from a list structure.
}
\author{
Aaron Taudt
}

