% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aneufinder-package.R
\docType{package}
\name{aneufinder-package}
\alias{aneufinder}
\alias{aneufinder-package}
\title{Copy-number detection in whole-genome single-cell sequencing data}
\description{
CNV detection in whole-genome single cell sequencing data using a Hidden Markov Model. The package implements CNV detection, commonly used plotting functions, export to BED format for upload to genome browsers, and measures for assessment of karyotype heterogeneity and quality metrics.
}
\details{
For a introduction to \pkg{aneufinder} see the vignette at \href{../doc/aneufinder.pdf}{A quick introduction to aneufinder}.
}
\author{
Aaron Taudt, David Porubsky
}

