% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aneufinder.R
\docType{package}
\name{aneufinder}
\alias{aneufinder}
\alias{aneufinder-package}
\title{Analysis of Aneuploidies and Copy Number Variation in Single-Cell-Sequencing Data}
\description{
TODO
}
\details{
TODO
}
\author{
Aaron Taudt, David Porubsky
}

