% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aneufinder.R
\name{Aneufinder}
\alias{Aneufinder}
\title{Wrapper function for the aneufinder package}
\usage{
Aneufinder(inputfolder, outputfolder, format, configfile = NULL, numCPU = 1,
  reuse.existing.files = TRUE, binsizes = 1e+06,
  variable.width.reference = NULL, reads.per.bin = NULL,
  pairedEndReads = FALSE, stepsize = NULL, assembly = NULL,
  chromosomes = NULL, remove.duplicate.reads = TRUE, min.mapq = 10,
  blacklist = NULL, correction.method = NULL, GC.BSgenome = NULL,
  mappability.reference = NULL, method = "univariate", eps = 0.1,
  max.time = 60, max.iter = 5000, num.trials = 15,
  states = c("zero-inflation", paste0(0:9, "-somy"), "+10-somy"),
  most.frequent.state.univariate = "2-somy",
  most.frequent.state.bivariate = "1-somy", resolution = c(3, 6),
  min.segwidth = 2, min.reads = 50, bw = 4 * binsizes[1], pval = 1e-08,
  cluster.plots = TRUE)
}
\arguments{
\item{inputfolder}{Folder with either BAM or BED files.}

\item{outputfolder}{Folder to output the results. If it does not exist it will be created.}

\item{format}{Either 'bam' or 'bed', depending if your \code{inputfolder} contains files in BAM or BED format.}

\item{configfile}{A file specifying the parameters of this function (without \code{inputfolder}, \code{outputfolder} and \code{configfile}). Having the parameters in a file can be handy if many samples with the same parameter settings are to be run. If a \code{configfile} is specified, it will take priority over the command line parameters.}

\item{numCPU}{The numbers of CPUs that are used. Should not be more than available on your machine.}

\item{reuse.existing.files}{A logical indicating whether or not existing files in \code{outputfolder} should be reused.}

\item{binsizes}{An integer vector with bin sizes. If more than one value is given, output files will be produced for each bin size.}

\item{variable.width.reference}{A BAM file that is used as reference to produce variable width bins. See \code{\link{variableWidthBins}} for details.}

\item{reads.per.bin}{Approximate number of desired reads per bin. The bin size will be selected accordingly. Output files are produced for each value.}

\item{pairedEndReads}{Set to \code{TRUE} if you have paired-end reads in your BAM files (not implemented for BED files).}

\item{stepsize}{Fraction of the binsize that the sliding window is offset at each step. Example: If \code{stepsize=0.1} and \code{binsizes=c(200000,500000)}, the actual stepsize in basepairs is 20000 and 50000, respectively.}

\item{assembly}{Please see \code{\link[GenomeInfoDb]{fetchExtendedChromInfoFromUCSC}} for available assemblies. Only necessary when importing BED files. BAM files are handled automatically. Alternatively a data.frame generated by \code{\link[GenomeInfoDb]{fetchExtendedChromInfoFromUCSC}}.}

\item{chromosomes}{If only a subset of the chromosomes should be imported, specify them here.}

\item{remove.duplicate.reads}{A logical indicating whether or not duplicate reads should be removed.}

\item{min.mapq}{Minimum mapping quality when importing from BAM files. Set \code{min.mapq=NULL} to keep all reads.}

\item{blacklist}{A \code{\link{GRanges}} or a bed(.gz) file with blacklisted regions. Reads falling into those regions will be discarded.}

\item{correction.method}{Correction methods to be used for the binned read counts. Currently any combination of \code{c('GC','mappability')}.}

\item{GC.BSgenome}{A \code{BSgenome} object which contains the DNA sequence that is used for the GC correction.}

\item{mappability.reference}{A file that serves as reference for mappability correction. Has to be the same format as specified by \code{format}.}

\item{method}{Any combination of \code{c('univariate','bivariate')}. Option \code{'univariate'} treats both strands as one, while option \code{'bivariate'} treats both strands separately. NOTE: SCEs can only be called when \code{method='bivariate'}.}

\item{eps}{Convergence threshold for the Baum-Welch algorithm.}

\item{max.time}{The maximum running time in seconds for the Baum-Welch algorithm. If this time is reached, the Baum-Welch will terminate after the current iteration finishes. The default -1 is no limit.}

\item{max.iter}{The maximum number of iterations for the Baum-Welch algorithm. The default -1 is no limit.}

\item{num.trials}{The number of trials to find a fit where state \code{most.frequent.state} is most frequent. Each time, the HMM is seeded with different random initial values.}

\item{states}{A subset or all of \code{c("zero-inflation","0-somy","1-somy","2-somy","3-somy","4-somy","+10-somy")}. This vector defines the states that are used in the Hidden Markov Model. The order of the entries should not be changed.}

\item{most.frequent.state.univariate}{One of the states that were given in \code{states}. The specified state is assumed to be the most frequent one when running the univariate HMM. This can help the fitting procedure to converge into the correct fit. Default is '2-somy'.}

\item{most.frequent.state.bivariate}{One of the states that were given in \code{states}. The specified state is assumed to be the most frequent one when running the bivariate HMM. This can help the fitting procedure to converge into the correct fit. Default is '1-somy'.}

\item{resolution}{An integer vector specifying the resolution at bin level at which to scan for SCE events.}

\item{min.segwidth}{Minimum segment length in bins when scanning for SCE events.}

\item{min.reads}{Minimum number of reads required for SCE refinement.}

\item{bw}{Bandwidth for SCE hotspot detection (see \code{\link{hotspotter}} for further details).}

\item{pval}{P-value for SCE hotspot detection (see \code{\link{hotspotter}} for further details).}

\item{cluster.plots}{A logical indicating whether plots should be clustered by similarity.}
}
\description{
This function is an easy-to-use wrapper to \link[aneufinder:binning]{bin the data}, \link[aneufinder:findCNVs]{find copy-number-variations}, \link[aneufinder:findSCEs]{find sister-chromatid-exchange} events, plot \link[aneufinder:heatmapGenomewide]{genomewide heatmaps}, \link[aneufinder:plot.aneuHMM]{distributions, profiles and karyograms}.
}
\examples{
\dontrun{
## The following call produces plots and genome browser files for all BAM files in "my-data-folder"
Aneufinder(inputfolder="my-data-folder", outputfolder="my-output-folder", format='bam')}

}
\author{
Aaron Taudt
}

