% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/importReads.R
\name{bed2GRanges}
\alias{bed2GRanges}
\title{Import BED file into GRanges}
\usage{
bed2GRanges(bedfile, assembly, chromosomes = NULL,
  remove.duplicate.reads = FALSE, min.mapq = 10,
  max.fragment.width = 1000)
}
\arguments{
\item{bedfile}{A file with aligned reads in BED format. The columns have to be c('chromosome','start','end','description','mapq','strand').}

\item{assembly}{Please see \code{\link[GenomeInfoDb]{fetchExtendedChromInfoFromUCSC}} for available assemblies. Only necessary when importing BED files. BAM files are handled automatically. Alternatively a data.frame generated by \code{\link[GenomeInfoDb]{fetchExtendedChromInfoFromUCSC}}.}

\item{chromosomes}{If only a subset of the chromosomes should be imported, specify them here.}

\item{remove.duplicate.reads}{A logical indicating whether or not duplicate reads should be removed.}

\item{min.mapq}{Minimum mapping quality when importing from BAM files. Set \code{min.mapq=NULL} to keep all reads.}

\item{max.fragment.width}{Maximum allowed fragment length. This is to filter out erroneously wrong fragments.}
}
\description{
Import aligned reads from a BED file into a \code{\link{GRanges}} object.
}

