% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/binning.R
\name{binning}
\alias{bam2binned}
\alias{bed2binned}
\alias{bedGraph2binned}
\alias{binning}
\title{Convert aligned reads from various file formats into read counts in equidistant bins}
\usage{
bedGraph2binned(bedGraphfile, chrom.length.file, outputfolder = "binned_data",
  binsizes = NULL, reads.per.bin = 10, numbins = NULL,
  chromosomes = NULL, GC.correction = TRUE, GC.correction.bsgenome,
  save.as.RData = TRUE, calc.complexity = TRUE,
  remove.duplicate.reads = TRUE, calc.spikyness = FALSE)

bam2binned(bamfile, bamindex = bamfile, pairedEndReads = FALSE,
  outputfolder = "binned_data", binsizes = NULL, reads.per.bin = 10,
  numbins = NULL, chromosomes = NULL, GC.correction = TRUE,
  GC.correction.bsgenome, save.as.RData = TRUE, calc.complexity = TRUE,
  remove.duplicate.reads = TRUE, calc.spikyness = FALSE)

bed2binned(bedfile, chrom.length.file, outputfolder = "binned_data",
  binsizes = NULL, reads.per.bin = 10, numbins = NULL,
  chromosomes = NULL, GC.correction = TRUE, GC.correction.bsgenome,
  save.as.RData = TRUE, calc.complexity = TRUE,
  remove.duplicate.reads = TRUE, calc.spikyness = FALSE)
}
\arguments{
\item{bedGraphfile}{A file in bedGraph format.}

\item{outputfolder}{Folder to which the binned data will be saved. If the specified folder does not exist, it will be created.}

\item{binsizes}{A vector with integer values which will be used for the binning. If more than one value is given, output files will be produced for each bin size.}

\item{reads.per.bin}{Approximate number of desired reads per bin. The bin size will be selected accordingly. Output files are produced for each value.}

\item{numbins}{Number of bins per chromosome. Each chromosome will have a different binsize! DO NOT USE UNLESS YOU KNOW WHAT YOU ARE DOING. Output files are produced for each value.}

\item{chromosomes}{If only a subset of the chromosomes should be binned, specify them here.}

\item{GC.correction}{Logical indicating whether or not to calculate GC-corrected reads.}

\item{GC.correction.bsgenome}{A \code{BSgenome} object that contains the DNA sequence that is used for the GC correction.}

\item{save.as.RData}{If set to \code{FALSE}, no output file will be written. Instead, a \link{GenomicRanges} object containing the binned data will be returned. Only the first binsize will be processed in this case.}

\item{calc.complexity}{A logical indicating whether or not to estimate library complexity.}

\item{remove.duplicate.reads}{A logical indicating whether or not duplicate reads should be removed.}

\item{calc.spikyness}{A logical indicating whether or not spikyness should be calculated.}

\item{bamfile}{A file in BAM format.}

\item{bamindex}{BAM index file. Can be specified without the .bai ending.}

\item{pairedEndReads}{Set to \code{TRUE} if you have paired-end reads in your file.}

\item{bedfile}{A file in BED format.}
}
\description{
Convert aligned reads in .bam or .bed format into read counts in equidistant windows. Convert signal values in .bedGraph format to signal counts in equidistant windows.
}
\details{
Convert aligned reads or signal values from various file formats into read counts in equidistant windows (bins). bam2binned() and bed2binned() use 'GenomicRanges::countOverlaps()' to calculate the read counts. Therefore, with small binsizes and large read lengths, one read fragment will often overlap more than one bin.
bedGraph2binned() sets the maximum signal value in a bin as value for that bin.
}
\section{Functions}{
\itemize{
\item \code{bedGraph2binned}: 

\item \code{bam2binned}: 

\item \code{bed2binned}: 
}}
\author{
Aaron Taudt
}

