% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualityMeasures.R
\name{clusterByQuality}
\alias{clusterByQuality}
\title{Cluster based on quality variables}
\usage{
clusterByQuality(hmms, G = 1:9, itmax = c(100, 100),
  measures = c("spikyness", "entropy", "num.segments", "bhattacharyya",
  "loglik"), orderBy = "spikyness", reverseOrder = FALSE)
}
\arguments{
\item{hmms}{A list of \code{\link{aneuHMM}} objects or a list of files that contain such objects.}

\item{G}{An integer vector specifying the number of clusters that are compared. See \code{\link[mclust:Mclust]{Mclust}} for details.}

\item{itmax}{The maximum number of outer and inner iterations for the \code{\link[mclust:Mclust]{Mclust}} function. See \code{\link[mclust:emControl]{emControl}} for details.}

\item{measures}{The quality measures that are used for the clustering. Supported is any combination of \code{c('spikyness','entropy','num.segments','bhattacharyya','loglik','complexity','avg.read.count','total.read.count','binsize')}.}

\item{orderBy}{The quality measure to order the clusters by. Default is \code{'spikyness'}.}

\item{reverseOrder}{Logical indicating whether the ordering by \code{orderBy} is reversed.}
}
\description{
This function uses the \pkg{\link{mclust}} package to cluster the input samples based on various quality measures.
}
\details{
The employed quality measures are:
\itemize{
\item Spikyness
\item Entropy
\item Number of segments
\item Bhattacharrya distance
}
}
\examples{
\dontrun{
## Get a list of HMMs
files <- list.files("your-outputfolder/results_univariate", full.names=TRUE)
cl <- clusterByQuality(files)
## Plot the clustering and print the parameters
plot(cl$Mclust, what='classification')
print(cl$parameters)
## Select files from the best 2 clusters for further processing
best.files <- unlist(cl$classification[1:2])}

}
\author{
Aaron Taudt
}

