% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterHMMs.R
\name{clusterHMMs}
\alias{clusterHMMs}
\title{Cluster objects}
\usage{
clusterHMMs(hmms, cluster = TRUE, classes = NULL, exclude.regions = NULL)
}
\arguments{
\item{hmms}{A list of \code{\link{aneuHMM}} or \code{\link{aneuBiHMM}} objects or a character vector of files that contains such objects.}

\item{cluster}{Either \code{TRUE} or \code{FALSE}, indicating whether the samples should be clustered by similarity in their CNV-state.}

\item{exclude.regions}{A \code{\link{GRanges-class}} with regions that will be excluded from the computation of the clustering. This can be useful to exclude regions with artifacts.}
}
\value{
An list() with ordered ID indices and the hierarchical clustering.
}
\description{
Cluster a list of \code{\link{aneuHMM}} or \code{\link{aneuBiHMM}} objects by similarity in their CNV-state.
}
\examples{
## Get results from a small-cell-lung-cancer
lung.folder <- system.file("extdata", "primary-lung", "hmms", package="AneuFinderData")
lung.files <- list.files(lung.folder, full.names=TRUE)
models <- loadFromFiles(lung.files)
\dontrun{
# Plot unclustered heatmap
heatmapGenomewide(models, cluster=FALSE)}
## Cluster and reorder the models
clust <- clusterHMMs(models)
models <- models[clust$IDorder]
\dontrun{
# Plot re-ordered heatmap
heatmapGenomewide(models, cluster=FALSE)}

}
