% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseBins.R
\name{collapseBins}
\alias{collapseBins}
\title{Collapse consecutive bins}
\usage{
collapseBins(data, column2collapseBy = NULL, columns2sumUp = NULL,
  columns2average = NULL, columns2getMax = NULL, columns2drop = NULL)
}
\arguments{
\item{data}{A data.frame containing the genomic coordinates in the first three columns.}

\item{column2collapseBy}{The number of the column which will be used to collapse all other inputs. If a set of consecutive bins has the same value in this column, they will be aggregated into one bin with adjusted genomic coordinates. If \code{NULL} directly adjacent bins will be collapsed.}

\item{columns2sumUp}{Column numbers that will be summed during the aggregation process.}

\item{columns2average}{Column numbers that will be averaged during the aggregation process.}

\item{columns2getMax}{Column numbers where the maximum will be chosen during the aggregation process.}

\item{columns2drop}{Column numbers that will be dropped after the aggregation process.}
}
\value{
A data.frame.
}
\description{
The function will collapse consecutive bins which have, for example, the same combinatorial state.
}
\details{
The following tables illustrate the principle of the collapsing:

Input data:
\tabular{rrrrrr}{
seqnames \tab start \tab end \tab column2collapseBy \tab moreColumns \tab columns2sumUp \cr
chr1     \tab     0 \tab 199 \tab                 2 \tab        1 10 \tab           1 3 \cr
chr1     \tab   200 \tab 399 \tab                 2 \tab        2 11 \tab           0 3 \cr
chr1     \tab   400 \tab 599 \tab                 2 \tab        3 12 \tab           1 3 \cr
chr1     \tab   600 \tab 799 \tab                 1 \tab        4 13 \tab           0 3 \cr
chr1     \tab   800 \tab 999 \tab                 1 \tab        5 14 \tab           1 3 \cr
}
Output data:
\tabular{rrrrrr}{
seqnames \tab start \tab end \tab column2collapseBy \tab moreColumns \tab columns2sumUp \cr
chr1     \tab     0 \tab 599 \tab                 2 \tab        1 10 \tab           2 9 \cr
chr1     \tab   600 \tab 999 \tab                 1 \tab        4 13 \tab           1 6 \cr
}
}
\examples{
## Get an example BED file with single-cell-sequencing reads
bedfile <- system.file("extdata", "KK150311_VI_07.bam.bed.gz", package="AneuFinderData")
## Bin the BAM file into bin size 1Mp
binned <- binReads(bedfile, assembly='mm10', binsize=1e6,
                  chromosomes=c(1:19,'X','Y'))
## Collapse the bins by chromosome and get average, summed and maximum read count
df <- as.data.frame(binned[[1]])
# Remove one bin for illustration purposes
df <- df[-3,]
head(df)
collapseBins(df, column2collapseBy='seqnames', columns2sumUp=c('width','counts'),
                       columns2average='counts', columns2getMax='counts',
                       columns2drop=c('mcounts','pcounts'))
collapseBins(df, column2collapseBy=NULL, columns2sumUp=c('width','counts'),
                       columns2average='counts', columns2getMax='counts',
                       columns2drop=c('mcounts','pcounts'))

}
\author{
Aaron Taudt
}

