% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{colors}
\alias{colors}
\alias{stateColors}
\alias{strandColors}
\title{\pkg{aneufinder} color scheme}
\usage{
stateColors(states = c("zero-inflation", paste0(0:9, "-somy"), "+10-somy",
  "total"))

strandColors(strands = c("+", "-"))
}
\arguments{
\item{states}{A character vector with states whose color should be returned.}

\item{strands}{A character vector with strands whose color should be returned. Any combination of \code{c('+','-','*')}.}
}
\value{
A character vector with colors.
}
\description{
Get the color schemes that are used in the aneufinder plots.
}
\section{Functions}{
\itemize{
\item \code{stateColors}: Colors that are used for the states.

\item \code{strandColors}: Colors that are used to distinguish strands.
}}
\examples{
## Make a nice pie chart with the aneufinder state color scheme
statecolors <- stateColors()
pie(rep(1,length(statecolors)), labels=names(statecolors), col=statecolors)

## Make a nice pie chart with the aneufinder strand color scheme
strandcolors <- strandColors()
pie(rep(1,length(strandcolors)), labels=names(strandcolors), col=strandcolors)

}

