% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctionMethods.R
\name{correctMappability}
\alias{correctMappability}
\title{Mappability correction}
\usage{
correctMappability(binned.data.list, same.binsize, reference, format, assembly,
  pairedEndReads = FALSE, min.mapq = 10, remove.duplicate.reads = TRUE,
  max.fragment.width = 1000)
}
\arguments{
\item{binned.data.list}{A \code{list} with \code{\link{binned.data}} objects or a list of filenames containing such objects.}

\item{same.binsize}{If \code{TRUE} the mappability correction will only be calculated once. Set this to \code{TRUE} if all \code{\link{binned.data}} objects describe the same genome at the same binsize.}

\item{reference}{A file or \code{\link{GRanges}} with aligned reads.}

\item{format}{Format of the reference, one of \code{c('bam','bed','GRanges')}.}

\item{assembly}{Please see \code{\link[GenomeInfoDb]{fetchExtendedChromInfoFromUCSC}} for available assemblies. Only necessary when importing BED files. BAM files are handled automatically. Alternatively a data.frame generated by \code{\link[GenomeInfoDb]{fetchExtendedChromInfoFromUCSC}}.}

\item{pairedEndReads}{Set to \code{TRUE} if you have paired-end reads in your BAM files (not implemented for BED files).}

\item{min.mapq}{Minimum mapping quality when importing from BAM files. Set \code{min.mapq=NULL} to keep all reads.}

\item{remove.duplicate.reads}{A logical indicating whether or not duplicate reads should be removed.}

\item{max.fragment.width}{Maximum allowed fragment length. This is to filter out erroneously wrong fragments due to mapping errors of paired end reads.}
}
\value{
A \code{list} with \code{\link{binned.data}} objects with adjusted read counts.
}
\description{
Correct a list of \code{\link{binned.data}} by mappability.
}
\author{
Aaron Taudt
}

