% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/findSCEs.R
\name{deltaWCalculator}
\alias{deltaWCalculator}
\title{Calculate deltaWs}
\usage{
deltaWCalculator(frags, reads.per.window = 100)
}
\arguments{
\item{frags}{A \code{\link{GRanges}} with read fragments (see \code{\link{bam2GRanges}}).}

\item{reads.per.window}{Number of reads in each dynamic window.}
}
\description{
This function will calculate differences between left and right window from a \code{\link{GRanges}} object with read fragments.
}
\author{
Aaron Taudt
}

