% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findCNVs.R
\name{DNAcopy.findCNVs}
\alias{DNAcopy.findCNVs}
\title{Find copy number variations (DNAcopy, univariate)}
\usage{
DNAcopy.findCNVs(binned.data, ID = NULL, CNgrid.start = 1.5, strand = "*")
}
\arguments{
\item{binned.data}{A \link{GRanges} object with binned read counts.}

\item{ID}{An identifier that will be used to identify this sample in various downstream functions. Could be the file name of the \code{binned.data} for example.}

\item{CNgrid.start}{Start parameter for the CNgrid variable. Very empiric. Set to 1.5 for normal data and 0.5 for Strand-seq data.}

\item{strand}{Find copy-numbers only for the specified strand. One of \code{c('+', '-', '*')}.}
}
\value{
An \code{\link{aneuHMM}} object.
}
\description{
\code{DNAcopy.findCNVs} classifies the binned read counts into several states which represent copy-number-variation.
}
