% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binReads.R
\name{estimateComplexity}
\alias{estimateComplexity}
\title{Estimate library complexity}
\usage{
estimateComplexity(reads)
}
\arguments{
\item{reads}{A \code{\link{GRanges}} object with read fragments. NOTE: Complexity estimation relies on duplicate reads and therefore the duplicates have to be present in the input.}
}
\value{
A \code{list} with estimated complexity values and plots.
}
\description{
Estimate library complexity using a very simple "Michaelis-Menten" approach and the sophisticated approach from the \pkg{\link{preseqR}} package.
}

