% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/binning.R
\name{estimateComplexity}
\alias{estimateComplexity}
\title{Estimate library complexity}
\usage{
estimateComplexity(reads)
}
\arguments{
\item{reads}{A \code{\link{GRanges}} object with read fragments. NOTE: Complexity estimation relies on duplicate reads and therefore the duplicates have to be present in the input.}
}
\description{
Estimate library complexity using a very simple "Michaelis-Menten" approach and the sophisticated approach from the \pkg{\link{preseqR}} package.
}

