% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export}
\alias{export}
\alias{exportCNVs}
\alias{exportGRanges}
\alias{exportReadCounts}
\title{Export genome browser viewable files}
\usage{
exportCNVs(hmms, filename, cluster = TRUE, export.CNV = TRUE,
  export.SCE = TRUE)

exportReadCounts(hmms, filename)

exportGRanges(gr, filename, header = TRUE, trackname = NULL, score = NULL,
  priority = NULL, append = FALSE, chromosome.format = "UCSC")
}
\arguments{
\item{hmms}{A list of \code{\link{aneuHMM}} objects or a character vector with files that contain such objects.}

\item{filename}{The name of the file that will be written. The appropriate ending will be appended, either ".bed.gz" for CNV-state or ".wig.gz" for read counts. Any existing file will be overwritten.}

\item{cluster}{If \code{TRUE}, the samples will be clustered by similarity in their CNV-state.}

\item{export.CNV}{A logical, indicating whether the CNV-state shall be exported.}

\item{export.SCE}{A logical, indicating whether breakpoints shall be exported.}

\item{gr}{A \code{\link{GRanges}} object.}

\item{header}{A logical indicating whether the output file will have a heading track line (\code{TRUE}) or not (\code{FALSE}).}

\item{trackname}{The name that will be used as track name and description in the header.}

\item{score}{A vector of the same length as \code{gr}, which will be used for the 'score' column in the BED file.}

\item{priority}{Priority of the track for display in the genome browser.}

\item{append}{Append to \code{filename}.}

\item{chromosome.format}{A character specifying the format of the chromosomes if \code{assembly} is specified. Either 'NCBI' for (1,2,3 ...) or 'UCSC' for (chr1,chr2,chr3 ...).#' @importFrom utils write.table}
}
\value{
\code{NULL}
}
\description{
Export copy-number-variation state or read counts as genome browser viewable file
}
\details{
Use \code{exportCNVs} to export the copy-number-variation state from an \code{\link{aneuHMM}} object in BED format.
Use \code{exportReadCounts} to export the binned read counts from an \code{\link{aneuHMM}} object in WIGGLE format.
Use \code{exportGRanges} to export a \code{\link{GRanges}} object in BED format.
}
\section{Functions}{
\itemize{
\item \code{exportCNVs}: Export CNV-state as .bed.gz file

\item \code{exportReadCounts}: Export binned read counts as .wig.gz file

\item \code{exportGRanges}: Export \code{\link{GRanges}} object as BED file.
}}
\examples{
\dontrun{
## Get results from a small-cell-lung-cancer
folder <- system.file("extdata", "primary-lung", "hmms", package="AneuFinderData")
files <- list.files(folder, full.names=TRUE)
## Export the CNV states for upload to the UCSC genome browser
exportCNVs(files, filename='upload-me-to-a-genome-browser', cluster=TRUE)}

}
\author{
Aaron Taudt
}

