% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSCEs.R
\name{filterSegments}
\alias{filterSegments}
\title{Filter segments by minimal size}
\usage{
filterSegments(segments, min.seg.width)
}
\arguments{
\item{segments}{A \code{\link{GRanges}} object.}

\item{min.seg.width}{The minimum segment width in base-pairs.}
}
\value{
The input \code{model} with adjusted segments.
}
\description{
\code{filterSegments} filters out segments below a specified minimal segment size. This can be useful to get rid of boundary effects from the Hidden Markov approach.
}
\examples{
## Load an HMM
file <- system.file("extdata/primary-lung/results_univariate/AvdB150303_I_001.bam.RData",
                    package="aneufinder")
hmm <- loadHmmsFromFiles(file)[[1]]
## Check number of segments before and after filtering
length(hmm$segments)
hmm$segments <- filterSegments(hmm$segments, min.seg.width=2*width(hmm$bins)[1])
length(hmm$segments)

}
\author{
Aaron Taudt
}

