% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBins.R
\name{fixedWidthBins}
\alias{fixedWidthBins}
\title{Make fixed-width bins}
\usage{
fixedWidthBins(bamfile = NULL, assembly = NULL, chrom.lengths = NULL,
  chromosome.format, binsizes = 1e+06, stepsizes = NULL,
  chromosomes = NULL)
}
\arguments{
\item{bamfile}{A BAM file from which the header is read to determine the chromosome lengths. If a \code{bamfile} is specified, option \code{assembly} is ignored.}

\item{assembly}{An assembly from which the chromosome lengths are determined. Please see \code{\link[GenomeInfoDb]{getChromInfoFromUCSC}} for available assemblies. This option is ignored if \code{bamfile} is specified. Alternatively a data.frame generated by \code{\link[GenomeInfoDb]{getChromInfoFromUCSC}}.}

\item{chrom.lengths}{A named character vector with chromosome lengths. Names correspond to chromosomes.}

\item{chromosome.format}{A character specifying the format of the chromosomes if \code{assembly} is specified. Either 'NCBI' for (1,2,3 ...) or 'UCSC' for (chr1,chr2,chr3 ...). If a \code{bamfile} or \code{chrom.lengths} is supplied, the format will be chosen automatically.}

\item{binsizes}{A vector of bin sizes in base pairs.}

\item{stepsizes}{A vector of step sizes in base pairs, the same length as \code{binsizes}.}

\item{chromosomes}{A subset of chromosomes for which the bins are generated.}
}
\value{
A \code{list()} of \code{\link{GRanges-class}} objects with fixed-width bins. If \code{stepsizes} is specified, a \code{list()} of \code{\link{GRangesList}} objects with one entry per step.
}
\description{
Make fixed-width bins based on given bin size.
}
\examples{
## Make fixed-width bins of size 500kb and 1Mb
bins <- fixedWidthBins(assembly='mm10', chromosome.format='NCBI', binsizes=c(5e5,1e6))
bins

}
\author{
Aaron Taudt
}
