% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBreakpoints.R
\name{getBreakpoints}
\alias{getBreakpoints}
\title{Extract breakpoints}
\usage{
getBreakpoints(model, fragments = NULL, confint = 0.99)
}
\arguments{
\item{model}{An \code{\link{aneuBiHMM}} object or a file that contains such an object.}

\item{fragments}{A \code{\link{GRanges}} object with read fragments or a file that contains such an object.}

\item{confint}{Desired confidence interval for breakpoints. Set \code{confint=NULL} to disable confidence interval estimation.}
}
\value{
A \code{\link{GRanges}} with breakpoint coordinates and confidence interals if \code{fragments} was specified.
}
\description{
Extract breakpoints with confidence intervals from an \code{\link{aneuBiHMM}} object.
}
\details{
Confidence intervals for breakpoints are estimated by going outwards from the breakpoint read by read, and performing a test of getting the observed or a more extreme outcome, given that the reads within the confidence interval belong to the other side of the breakpoint.
}
\examples{
## Get an example BED file with single-cell-sequencing reads
bedfile <- system.file("extdata", "KK150311_VI_07.bam.bed.gz", package="AneuFinderData")
## Bin the data into bin size 1Mp
readfragments <- binReads(bedfile, assembly='mm10', binsize=1e6,
                  chromosomes=c(1:19,'X','Y'), reads.return=TRUE)
binned <- binReads(bedfile, assembly='mm10', binsize=1e6,
                  chromosomes=c(1:19,'X','Y'))
## Fit the Hidden Markov Model
model <- findCNVs.strandseq(binned[[1]], eps=0.01, max.time=60)
## Add confidence intervals
breakpoints <- getBreakpoints(model, readfragments)

}
