% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualityMeasures.R
\name{getQC}
\alias{getQC}
\title{Obtain a data.frame with quality metrics}
\usage{
getQC(models)
}
\arguments{
\item{models}{A list of \code{\link{GRanges}} or \code{\link{aneuHMM}} objects or a list of files that contain such objects.}
}
\value{
A data.frame with columns
}
\description{
Obtain a data.frame with quality metrics from a list of \code{\link{aneuHMM}} objects or a list of files that contain such objects.
}
\examples{
## Get a list of HMMs
folder <- system.file("extdata", "primary-lung", "hmms", package="AneuFinderData")
files <- list.files(folder, full.names=TRUE)
df <- getQC(files)
}
\author{
Aaron Taudt
}

