% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/findSCEs.R
\name{getSCEcoordinates}
\alias{getSCEcoordinates}
\title{Get SCE coordinates}
\usage{
getSCEcoordinates(model, resolution = c(3, 6))
}
\arguments{
\item{model}{An \code{\link{aneuBiHMM}} object.}

\item{resolution}{An integer vector specifying the number of bins for filtering. Segments below these numbers will be filtered out for SCE determination.}
}
\value{
A \code{\link{GRanges}} object containing the SCE coordinates.
}
\description{
Extracts the coordinates of a sister chromatid exchanges (SCE) from an \code{\link{aneuBiHMM}} object.
}
\author{
Aaron Taudt
}

