% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getSegments.R
\name{getSegments}
\alias{getSegments}
\title{Extract segments and cluster}
\usage{
getSegments(hmm.list, cluster = TRUE)
}
\arguments{
\item{hmm.list}{A list of \code{\link{aneuHMM}} or \code{\link{aneuBiHMM}} objects or files that contain such objects.}

\item{cluster}{Either \code{TRUE} or \code{FALSE}, indicating whether the samples should be clustered by similarity in their CNV-state.}
}
\value{
A \code{list()} with (clustered) segments and SCE coordinates.
}
\description{
Extract segments and ID from a list of \code{\link{aneuHMM}} or \code{\link{aneuBiHMM}} objects and cluster if desired.
}

