% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{heatmapAneuploidies}
\alias{heatmapAneuploidies}
\title{Plot aneuploidy state}
\usage{
heatmapAneuploidies(hmms, ylabels = NULL, cluster = TRUE,
  as.data.frame = FALSE)
}
\arguments{
\item{hmms}{A list of \code{\link{aneuHMM}} objects or files that contain such objects.}

\item{ylabels}{A vector with labels for the y-axis. The vector must have the same length as \code{hmms}. If \code{NULL} the IDs from the \code{\link{aneuHMM}} objects will be used.}

\item{cluster}{If \code{TRUE}, the samples will be clustered by similarity in their CNV-state.}

\item{as.data.frame}{If \code{TRUE}, instead of a plot, a data.frame with the aneuploidy state for each sample will be returned.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot}} object or a data.frame, depending on option \code{as.data.frame}.
}
\description{
Plot a heatmap of aneuploidy state for multiple samples. Samples can be clustered and the output can be returned as data.frame.
}
\author{
Aaron Taudt
}

