% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{heatmapGenomewide}
\alias{heatmapGenomewide}
\title{Genome wide heatmap of CNV-state}
\usage{
heatmapGenomewide(hmms, ylabels = NULL, classes = NULL,
  reorder.by.class = TRUE, classes.color = NULL, file = NULL,
  cluster = TRUE, plot.breakpoints = TRUE, hotspots = NULL,
  exclude.regions = NULL)
}
\arguments{
\item{hmms}{A list of \code{\link{aneuHMM}} objects or a character vector with files that contain such objects.}

\item{ylabels}{A vector with labels for the y-axis. The vector must have the same length as \code{hmms}. If \code{NULL} the IDs from the \code{\link{aneuHMM}} objects will be used.}

\item{classes}{A character vector with the classification of the elements on the y-axis. The vector must have the same length as \code{hmms}.}

\item{reorder.by.class}{If \code{TRUE}, the dendrogram will be reordered to display similar classes next to each other.}

\item{classes.color}{A (named) vector with colors that are used to distinguish \code{classes}. Names must correspond to the unique elements in \code{classes}.}

\item{file}{A PDF file to which the heatmap will be plotted.}

\item{cluster}{Either \code{TRUE} or \code{FALSE}, indicating whether the samples should be clustered by similarity in their CNV-state.}

\item{plot.breakpoints}{Logical indicating whether breakpoints should be plotted.}

\item{hotspots}{A \code{\link{GRanges}} object with coordinates of genomic hotspots (see \code{\link{hotspotter}}).}

\item{exclude.regions}{A \code{\link{GRanges}} with regions that will be excluded from the computation of the clustering. This can be useful to exclude regions with artifacts.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot}} object or \code{NULL} if a file was specified.
}
\description{
Plot a genome wide heatmap of copy number variation state. This heatmap is best plotted to file, because in most cases it will be too big for cleanly plotting it to screen.
}
\examples{
## Get results from a small-cell-lung-cancer
lung.folder <- system.file("extdata", "primary-lung", "hmms", package="AneuFinderData")
lung.files <- list.files(lung.folder, full.names=TRUE)
## Get results from the liver metastasis of the same patient
liver.folder <- system.file("extdata", "metastasis-liver", "hmms", package="AneuFinderData")
liver.files <- list.files(liver.folder, full.names=TRUE)
## Plot a clustered heatmap
classes <- c(rep('lung', length(lung.files)), rep('liver', length(liver.files)))
labels <- c(paste('lung',1:length(lung.files)), paste('liver',1:length(liver.files)))
heatmapGenomewide(c(lung.files, liver.files), ylabels=labels, classes=classes,
                 classes.color=c('blue','red'))

}

