% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{heatmapGenomewideClusters}
\alias{heatmapGenomewideClusters}
\title{Plot heatmaps for quality control}
\usage{
heatmapGenomewideClusters(cl = NULL, cutree = NULL, file = NULL, ...)
}
\arguments{
\item{cl}{The return value of \code{\link{clusterByQuality}}.}

\item{cutree}{The return value of \code{\link[stats]{cutree}}, where the names correspond to the filenames to be loaded.}

\item{file}{A character specifying the output file.}

\item{...}{Further parameters passed on to \code{\link{heatmapGenomewide}}.}
}
\value{
A \code{\link[cowplot]{cowplot}} object or \code{NULL} if a file was specified.
}
\description{
This function is a convenient wrapper to call \code{\link{heatmapGenomewide}} for all clusters after calling \code{\link{clusterByQuality}} and plot the heatmaps into one pdf for efficient comparison.
}
\examples{
## Get a list of HMMs and cluster them
folder <- system.file("extdata", "primary-lung", "hmms", package="AneuFinderData")
files <- list.files(folder, full.names=TRUE)
cl <- clusterByQuality(files, G=5)
heatmapGenomewideClusters(cl=cl)

## Plot sub-clones of the largest cluster
largest.cluster <- which.max(sapply(cl$classification, length))
files <- cl$classification[[largest.cluster]]
clust <- clusterHMMs(files)
groups <- cutree(tree = clust$hclust, k = 5)
heatmapGenomewideClusters(cutree = groups, cluster = FALSE)

}
