% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedimport.R
\name{importBed}
\alias{importBed}
\title{Read bed-file into GRanges}
\usage{
importBed(bedfile, skip = 0, chromosome.format = "NCBI")
}
\arguments{
\item{bedfile}{Filename of the bed or bed.gz file.}

\item{skip}{Number of lines to skip at the beginning.}

\item{chromosome.format}{Desired format of the chromosomes. Either 'NCBI' for (1,2,3 ...) or 'UCSC' for (chr1,chr2,chr3 ...).}
}
\value{
A \code{\link{GRanges}} object with the contents of the bed-file.
}
\description{
This is a simple convenience function to read a bed(.gz)-file into a \code{\link{GRanges}} object. The bed-file is expected to have the following fields: \code{chromosome, start, end, name, score, strand}.
}
\examples{
## Get an example BED file with single-cell-sequencing reads
bedfile <- system.file("extdata/KK150311-VI_07.bam.bed.gz", package="aneufinder")
## Import the file and skip the first 10 lines
data <- importBed(bedfile, skip=10)

}
\author{
Aaron Taudt
}

