% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadModelsFromFiles.R
\name{loadGRangesFromFiles}
\alias{loadGRangesFromFiles}
\title{Load GRanges from files}
\usage{
loadGRangesFromFiles(files)
}
\arguments{
\item{files}{A list of files that contain \code{\link{GRanges}} objects.}
}
\value{
A list() containing all loaded \code{\link{GRanges}} objects.
}
\description{
Load \code{\link{GRanges}} objects from file into a list.
}
\examples{
\dontrun{
## Get some files that you want to load
files <- list.files("folder-with-saved-GRanges-RData-files",
                   full.names=TRUE)
grlist <- loadGRangesFromFiles(files)
## Plot one of them
plot(grlist[[1]]}

}
\author{
Aaron Taudt
}

