\name{loadHmmsFromFiles}
\alias{loadHmmsFromFiles}

\title{
Easy loading of multiple \code{\link{hmm}} files
}
\description{
This function will load all specified \code{\link{hmm}}s.
}
\usage{
loadHmmsFromFiles(hmm.list)
}
\arguments{
\item{hmm.list}{A list of files that contain \code{\link{hmm}}s.}
}

\value{
A list containing all the specified \code{\link{hmm}}s is returned.
}
\author{Aaron Taudt}
\examples{
}
