% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadModelsFromFiles.R
\name{loadHmmsFromFiles}
\alias{loadHmmsFromFiles}
\title{Load HMMs from files}
\usage{
loadHmmsFromFiles(hmms, strict = FALSE)
}
\arguments{
\item{hmms}{A list of files that contain \code{\link{aneuHMM}} objects.}

\item{strict}{If any of the loaded objects is not a \code{\link{aneuHMM}} object, an error (\code{strict=TRUE}) or a warning (\code{strict=FALSE}) will be generated.}
}
\value{
A list() containing all loaded \code{\link{aneuHMM}} objects.
}
\description{
Load \code{\link{aneuHMM}} objects from file into a list.
}
\author{
Aaron Taudt
}

