% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.aneuHMM}
\alias{plot.aneuHMM}
\title{Plotting function for \code{\link{aneuHMM}} objects}
\usage{
\method{plot}{aneuHMM}(x, type = "karyogram", ...)
}
\arguments{
\item{x}{An \code{\link{aneuHMM}} object.}

\item{type}{Type of the plot, one of \code{c('arrayCGH', 'histogram', 'karyogram')}. You can also specify the type with an integer number.
\describe{
  \item{\code{karyogram}}{A karyogram-like chromosome overview with CNV-state.}
  \item{\code{histogram}}{A histogram of binned read counts with fitted mixture distribution.}
  \item{\code{karyogram}}{An arrayCGH-like chromosome overview with CNV-state.}
}}

\item{...}{Additional arguments for the different plot types.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot}} object.
}
\description{
Make different types of plots for \code{\link{aneuHMM}} objects.
}

