% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.GRanges}
\alias{plot.GRanges}
\title{Plotting function for binned read counts}
\usage{
\method{plot}{GRanges}(x, chromosome = NULL, start = NULL, end = NULL,
  ...)
}
\arguments{
\item{x}{An \code{\link{aneuHMM}} object.}

\item{chromosome}{Plot the histogram only for the specified chromosome, start and end position.}

\item{start}{Plot the histogram only for the specified chromosome, start and end position.}

\item{end}{Plot the histogram only for the specified chromosome, start and end position.}

\item{...}{Additional arguments not implemented.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot}} object.
}
\description{
Make plots for binned read counts from \code{\link{binned.data}}
}

