% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotArray}
\alias{plotArray}
\title{ArrayCGH-like chromosome overview}
\usage{
plotArray(model, both.strands = FALSE, plot.SCE = TRUE, file = NULL)
}
\arguments{
\item{model}{A \code{\link{aneuHMM}} object or \code{\link{binned.data}}.}

\item{file}{A PDF file where the plot will be saved.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot}} object or \code{NULL} if a file was specified.
}
\description{
Plot an arrayCGH-like chromosome overview with read counts and CNV-state from a \code{\link{aneuHMM}} object or \code{\link{binned.data}}.
}

