% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotHeterogeneity}
\alias{plotHeterogeneity}
\title{Heterogeneity vs. Aneuploidy}
\usage{
plotHeterogeneity(hmms, hmms.list = NULL, normalChromosomeNumbers = NULL,
  plot = TRUE)
}
\arguments{
\item{hmms}{A list of \code{\link{aneuHMM}} objects or files that contain such objects.}

\item{hmms.list}{A named list() of lists of \code{\link{aneuHMM}} objects or files that contain such objects.}

\item{normalChromosomeNumbers}{A named integer vector or matrix with physiological copy numbers, where each element (vector) or column (matrix) corresponds to a chromosome. This is useful to specify male or female samples, e.g. \code{c('X'=2)} for female samples or \code{c('X'=1,'Y'=1)} for male samples. Specify a vector if all your \code{hmms} have the same physiological copy numbers. Specify a matrix if your \code{hmms} have different physiological copy numbers (e.g. a mix of male and female samples). If not specified otherwise, '2' will be assumed for all chromosomes.}

\item{plot}{A logical indicating whether to plot or to return the underlying data.frame.}
}
\value{
A \code{\link[ggplot2]{ggplot}} object or a data.frame if \code{plot=FALSE}.
}
\description{
Make heterogeneity vs. aneuploidy plots using individual chromosomes as datapoints.
}
\examples{
## Get results from a small-cell-lung-cancer
lung.folder <- system.file("extdata", "primary-lung", "hmms", package="AneuFinderData")
lung.files <- list.files(lung.folder, full.names=TRUE)
## Get results from the liver metastasis of the same patient
liver.folder <- system.file("extdata", "metastasis-liver", "hmms", package="AneuFinderData")
liver.files <- list.files(liver.folder, full.names=TRUE)
## Make heterogeneity plots
plotHeterogeneity(hmms.list = list(lung=lung.files, liver=liver.files))

}

