% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotProfile}
\alias{plotProfile}
\title{Read count and CNV profile}
\usage{
plotProfile(model, both.strands = FALSE, plot.breakpoints = FALSE,
  file = NULL, normalize.counts = NULL)
}
\arguments{
\item{model}{A \code{\link{aneuHMM}} object or \code{\link{binned.data}}.}

\item{both.strands}{If \code{TRUE}, strands will be plotted separately.}

\item{plot.breakpoints}{Logical indicating whether breakpoints should be plotted.}

\item{file}{A PDF file where the plot will be saved.}

\item{normalize.counts}{An character giving the copy number state to which to normalize the counts, e.g. '1-somy', '2-somy' etc.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot}} object or \code{NULL} if a file was specified.
}
\description{
Plot a profile with read counts and CNV-state from a \code{\link{aneuHMM}} object or \code{\link{binned.data}}.
}

