% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qualityMeasures.R
\name{qualityControl}
\alias{qc.entropy}
\alias{qc.spikyness}
\alias{qualityControl}
\title{Quality control measures for binned read counts}
\usage{
qc.spikyness(reads)

qc.entropy(reads)
}
\arguments{
\item{reads}{A vector of binned read counts.}
}
\description{
Calculate various quality control measures on binned read counts.
}
\details{
The Shannon entropy is defined as
\eqn{S = - sum( n * log(n) )}, where \eqn{n = reads/sum(reads)}.\cr\cr
Spikyness is defined as \eqn{K = sum(abs(diff(reads))) / sum(reads)}.
}
\section{Functions}{
\itemize{
\item \code{qc.spikyness}: Calculate the spikyness of a library

\item \code{qc.entropy}: Calculate the Shannon entropy of a library
}}
\author{
Aaron Taudt
}

