% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refine_changepoints.R
\name{subsetByCNVprofile}
\alias{subsetByCNVprofile}
\title{Get IDs of a subset of models}
\usage{
subsetByCNVprofile(hmms, profile)
}
\arguments{
\item{hmms}{A list of \code{\link{aneuHMM}} objects or files that contain such objects.}

\item{profile}{A \code{\link{GRanges}} object with metadata column 'expected.state' and optionally columns 'expected.mstate' and 'expected.pstate'.}
}
\value{
A named logical vector with \code{TRUE} for all models that are concordant with the given \code{profile}.
}
\description{
Get the IDs of models that have a certain CNV profile. The result will be \code{TRUE} for models that overlap all specified ranges in \code{profile} by at least one base pair with the correct state.
}
\examples{
## Get results from a small-cell-lung-cancer
lung.folder <- system.file("extdata/primary-lung/results_univariate", package="aneufinder")
lung.files <- list.files(lung.folder, full.names=TRUE)
## Get all files that have a trisomy on chromosome 1 and tetrasomy on chromosome 2
profile <- GRanges(seqnames=c('1','2'), ranges=IRanges(start=c(1,1), end=c(195471971,182113224)),
                  expected.state=c('trisomy','tetrasomy'))
ids <- subsetByCNVprofile(lung.files, profile)
print(which(ids))

}

