% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/variableWidthBins.R
\name{variableWidthBins}
\alias{variableWidthBins}
\title{Make variably sized bins}
\usage{
variableWidthBins(bamfile, bamindex = bamfile, binsizes, chromosomes = NULL,
  pairedEndReads = FALSE, remove.duplicate.reads = FALSE, min.mapq = 10,
  max.fragment.width = 1000)
}
\arguments{
\item{bamfile}{A sorted BAM file.}

\item{bamindex}{BAM index file. Can be specified without the .bai ending. If the index file does not exist it will be created and a warning is issued.}

\item{binsizes}{A vector with binsizes. Resulting bins will be close to the specified binsizes.}

\item{chromosomes}{If only a subset of the chromosomes should be imported, specify them here.}

\item{pairedEndReads}{Set to \code{TRUE} if you have paired-end reads in your file.}

\item{remove.duplicate.reads}{A logical indicating whether or not duplicate reads should be removed.}

\item{min.mapq}{Minimum mapping quality when importing from BAM files. Set \code{min.mapq=NULL} to keep all reads.}

\item{max.fragment.width}{Maximum allowed fragment length. This is to filter out erroneously wrong fragments due to mapping errors of paired end reads.}
}
\value{
A \code{list} with one \code{\link{GRanges}} object for each element in \code{binsizes}.
}
\description{
Make variable-width bins based on a reference BAM file. This can be a simulated file (produced by \code{\link{simulateReads}} and aligned with your favourite aligner) or a real reference.
}
\details{
Variable-width bins are produced by first binning the reference BAM file with fixed-width bins and selecting the desired number of reads per bin as the (non-zero) maximum of the histogram. A new set of bins is then generated such that every bin contains the desired number of reads.
}
\author{
Aaron Taudt
}

