% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rwConfig.R
\name{writeConfig}
\alias{writeConfig}
\title{Write aneufinder configuration file}
\usage{
writeConfig(conf, configfile)
}
\arguments{
\item{conf}{A list structure with parameter values. Each entry will be written in one line.}

\item{configfile}{Filename of the outputfile.}
}
\description{
Write an aneufinder configuration file from a list structure.
}
\author{
Aaron Taudt
}

