% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinbinom.R
\name{zinbinom}
\alias{zinbinom}
\alias{dzinbinom}
\alias{pzinbinom}
\alias{qzinbinom}
\alias{rzinbinom}
\title{The Zero-inflated Negative Binomial Distribution}
\usage{
dzinbinom(x, w, size, prob, mu)

pzinbinom(q, w, size, prob, mu, lower.tail = TRUE)

qzinbinom(p, w, size, prob, mu, lower.tail = TRUE)

rzinbinom(n, w, size, prob, mu)
}
\arguments{
\item{x}{Vector of (non-negative integer) quantiles.}

\item{w}{Weight of the zero-inflation. \code{0 <= w <= 1}.}

\item{size}{Target for number of successful trials, or dispersion parameter (the shape parameter of the gamma mixing distribution). Must be strictly positive, need not be integer.}

\item{prob}{Probability of success in each trial. \code{0 < prob <= 1}.}

\item{mu}{Alternative parametrization via mean: see \sQuote{Details}.}

\item{q}{Vector of quantiles.}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P[X \le x]}, otherwise, \eqn{P[X > x]}.}

\item{p}{Vector of probabilities.}

\item{n}{number of observations. If \code{length(n) > 1}, the length is taken to be the number required.}
}
\value{
dzinbinom gives the density, pzinbinom gives the distribution function, qzinbinom gives the quantile function, and rzinbinom generates random deviates.
}
\description{
Density, distribution function, quantile function and random
generation for the zero-inflated negative binomial distribution with parameters
\code{w}, \code{size} and \code{prob}.
}
\details{
The zero-inflated negative binomial distribution with \code{size} \eqn{= n} and
 \code{prob} \eqn{= p} has density
 \deqn{
   p(x) = w + (1-w) \frac{\Gamma(x+n)}{\Gamma(n) x!} p^n (1-p)^x}{
   w + (1-w) * \Gamma(x+n)/(\Gamma(n) x!) p^n (1-p)^x}
 for \eqn{x = 0}, \eqn{n > 0}, \eqn{0 < p \le 1} and \eqn{0 \le w \le 1}.

\deqn{
   p(x) = (1-w) \frac{\Gamma(x+n)}{\Gamma(n) x!} p^n (1-p)^x}{
   (1-w) * \Gamma(x+n)/(\Gamma(n) x!) p^n (1-p)^x}
 for \eqn{x = 1, 2, \ldots}, \eqn{n > 0}, \eqn{0 < p \le 1} and \eqn{0 \le w \le 1}.
}
\section{Functions}{
\itemize{
\item \code{dzinbinom}: gives the density

\item \code{pzinbinom}: gives the cumulative distribution function

\item \code{qzinbinom}: gives the quantile function

\item \code{rzinbinom}: random number generation
}}

\seealso{
\link{Distributions} for standard distributions, including
 \code{\link{dbinom}} for the binomial, \code{\link{dnbinom}} for the negative binomial, \code{\link{dpois}} for the
 Poisson and \code{\link{dgeom}} for the geometric distribution, which
 is a special case of the negative binomial.
}
\author{
Matthias Heinig, Aaron Taudt
}
